update to zig 0.13.0, compile warning fix

+++ blockset/code/MemoryEmulation.c
@@ -74,7 +74,7 @@ void MemoryEmulation_WriteToMemory(char* key, float value, float* oldValue)
 		float dummy = (float) 0xffffffff;
 		MemoryEmulation_ReadFromMemory(key,&dummy);
 		/* If dummy is changed, the value is already there so ignore init value */
-		if(dummy != 0xffffffff)
+		if(dummy != (float) 0xffffffff)
 		{
 		return;
 		}
